+++ parsing/parsing_utils.c
@@ -6,7 +6,7 @@
 /*   By: fhanuise <fhanuise@student.42belgium.be    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/08 14:06:10 by fhanuise          #+#    #+#             */
-/*   Updated: 2026/04/30 13:44:21 by fhanuise         ###   ########.fr       */
+/*   Updated: 2026/04/30 14:35:48 by fhanuise         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -86,7 +86,7 @@ char	**alloc_map(char *file, int start, int count)
 	return (map);
 }
 
-unsigned long createRGB(int r, int g, int b)
+int	 createRGB(int r, int g, int b)
 {   
     return ((r & 0xff) << 16) + ((g & 0xff) << 8) + (b & 0xff);
 }
